refactor(GiniCaptureSDK): refactor error type for error screen
PIA-3333
